Fixing issue #1
With this commit i am fixing the issue that some insert tags
don't have a page context. Now you can define insert tags globaly
as well. They are now in the system settings in the globals section.

But you still can override them in the page tree as normal. The order is

* Global Tags
* Root Page
* Normal pages

Signed-off-by: Leo Unglaub <leo.unglaub@2create.at>